requirements.txt
numpy
scipy
pydantic
pyyaml
pytest
